refactor: consolidate formatting functions into src/lib/format.ts and update imports
