+++ src/app/analyze/virtualized-table/virtualized-table-header.tsx
@@ -37,8 +37,6 @@ export const VirtualizedTableHeader = <TData,>({
           {virtualColumns.map((virtualColumn) => {
             const header = headerGroup.headers[virtualColumn.index];
             if (!header) return null;
-            console.log('header size', header.getSize());
-            console.log('header id', virtualColumn.size);
             return (
               <TableHead
                 data-testid={`header-cell-${header.id}`}
